requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.26.56
botocore==1.29.56
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.1
daphne==4.0.0
Django==4.1.5
django-celery-beat==2.4.0
django-storages==1.13.2
django-timezone-field==5.0
exceptiongroup==1.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
msgpack==1.0.4
numpy==1.24.1
opencv-python==4.7.0.68
outcome==1.2.0
pandas==1.5.2
Pillow==9.4.0
prompt-toolkit==3.0.36
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
s3transfer==0.6.0
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.6
whitenoise==6.4.0
wsproto==1.2.0
zope.interface==5.5.2
